added trickle charge state
the final charging requires the dc2dc to be turned off to be able to properly measure the the current drop. the trickle charging state is the same as the charging state but turns off the dc2dc and is only entered once the highest cell reaches a specific high voltage.