config parameters rework
Added functionality to set config parameters which saves them to the eeprom automatically. When the BMS loads, it loads the config from the EEPROM automatically and verifies the magic bytes and the version are matching. If not matching, it reloads default values. When setting a new param, it checks it within range.